requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
scikit-learn
xmltodict
PyYAML
tqdm
tensorboard
pandas
seaborn